Added an old schema change

+++ t/test-genome-DBs/multi/compara/table.sql
@@ -510,7 +510,7 @@ CREATE TABLE `other_member_sequence` (
 ) ENGINE=MyISAM DEFAULT CHARSET=latin1 MAX_ROWS=10000000 AVG_ROW_LENGTH=60000;
 
 CREATE TABLE `peptide_align_feature` (
-  `peptide_align_feature_id` int(10) unsigned NOT NULL AUTO_INCREMENT,
+  `peptide_align_feature_id` bigint(20) unsigned NOT NULL AUTO_INCREMENT,
   `qmember_id` int(10) unsigned NOT NULL,
   `hmember_id` int(10) unsigned NOT NULL,
   `qgenome_db_id` int(10) unsigned DEFAULT NULL,
